requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.2
google-api-core==2.1.0
google-auth==2.2.1
google-cloud-core==2.1.0
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
h11==0.12.0
idna==3.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
